Moved operator translation from commands.py to each assembler, as it should be. Created a separate source file for Cpu. Made a class for a move definition and route all address conversions from binary to runtime or vice-versa through that. Created a new source file for the Subroutine class. Comments listing reference addresses are output with their move_id where needed.
