+++ examples/go
@@ -126,8 +126,8 @@ test_8080() {
 test "moving1.py" "moving.orig"
 test "moving2.py" "moving.orig"
 test "moving3.py" "moving.orig"
-#test_8080 "example8080.py" "example8080.orig"
-#test_8080 "test8080.py" "test8080.orig"
+test_8080 "example8080.py" "example8080.orig"
+test_8080 "test8080.py" "test8080.orig"
 test "helloworld.py" "helloworld.orig"
 test "local_labels.py" "local_labels.orig"
 test "basic4.py" "basic4.orig"
