+++ py8dis/movemanager.py
@@ -23,22 +23,23 @@ import contextlib
 import config
 import utils
 import memorymanager
+from movedefinition import MoveDefinition
 from memorymanager import BinaryLocation, RuntimeLocation, RuntimeAddr, BinaryAddr
 
 # active_move_ids is a list of MoveId objects. It is managed by the
 # MoveId class itself.
-base_move_id    = 0     # Integer value for the first move id.
+BASE_MOVE_ID    = 0     # Integer value for the first move id.
 active_move_ids = []
 
 # move_definitions is a list of moves, each of the form (dest, source, len)
 # It starts as a single move encompassing the entire memory map. Later
 # moves "steal" binary addresses from earlier moves.
-move_definitions = [(RuntimeAddr(0), BinaryAddr(0), 0x10000)]
+move_definitions = [MoveDefinition(RuntimeAddr(0), BinaryAddr(0), 0x10000)]
 
 def is_valid_move_id(move_id):
     """Sanity check that the move_id is within the expected range"""
 
-    return move_id == base_move_id or 0 <= move_id < len(move_definitions)
+    return move_id == BASE_MOVE_ID or 0 <= move_id < len(move_definitions)
 
 class MoveId(int):
     """Class for a move id, that via scope manages a list of active
@@ -58,7 +59,7 @@ class MoveId(int):
         active_move_ids.pop()
 
 # Create array of 65536 instances of one initial MoveId object
-move_id_for_binary_addr = [MoveId(base_move_id)] * 0x10000
+move_id_for_binary_addr = [MoveId(BASE_MOVE_ID)] * 0x10000
 
 def add_move(dest, source, length):
     """Register the existence of a 'move' at runtime that relocates memory
@@ -78,7 +79,7 @@ def add_move(dest, source, length):
     assert length > 0
 
     # Create a new move
-    move_definitions.append((dest, source, length))
+    move_definitions.append(MoveDefinition(dest, source, length))
 
     # Create a new move id
     move_id = MoveId(len(move_definitions) - 1)
@@ -97,10 +98,12 @@ def is_valid_runtime_addr_for_move_id(runtime_addr, move_id):
     assert memorymanager.is_valid_runtime_addr(runtime_addr)
     assert is_valid_move_id(move_id)
 
+    # Find the move definition
     md = move_definitions[move_id]
+
     # We *include* the end address here; this accounts for the fact that a label can be
     # defined after the last classification in a move.
-    return md[0] <= runtime_addr <= (md[0] + md[2])
+    return md.is_in_move_dest(runtime_addr, include_end_address=True)
 
 def b2r(binary_addr):
     """
@@ -116,12 +119,8 @@ def b2r(binary_addr):
 
     # Gather information about the move
     move_id = move_id_for_binary_addr[binary_addr]
-    move_dest, move_source, move_length = move_definitions[move_id]
-
-    assert move_source <= binary_addr < (move_source + move_length)
-    return RuntimeAddr(move_dest + (binary_addr - move_source))
-
-
+    md = move_definitions[move_id]
+    return md.convert_binary_to_runtime_addr(binary_addr)
 
 # `cache` holds a list of move_ids for each runtime address.
 cache = None
@@ -144,21 +143,21 @@ def move_ids_for_runtime_addr(runtime_addr):
     if cache_move_definitions_len is None or len(move_definitions) != cache_move_definitions_len:
         cache = collections.defaultdict(set)
         for binary_addr, move_id in enumerate(move_id_for_binary_addr):
-            if move_id != base_move_id: # TODO: special case feels a bit awkward
+            if move_id != BASE_MOVE_ID: # TODO: special case feels a bit awkward
                 cache[b2r(BinaryAddr(binary_addr))].add(move_id)
         cache_move_definitions_len = len(move_definitions)
     return cache[runtime_addr]
 
 def make_binloc(binary_loc):
     if utils.is_integer_type(binary_loc):
-        last_move_id = active_move_ids[-1] if active_move_ids else base_move_id
+        last_move_id = active_move_ids[-1] if active_move_ids else BASE_MOVE_ID
         return BinaryLocation(BinaryAddr(binary_loc), last_move_id)
     assert isinstance(binary_loc, BinaryLocation)
     return binary_loc
 
 def make_runloc(runtime_loc):
     if utils.is_integer_type(runtime_loc):
-        last_move_id = active_move_ids[-1] if active_move_ids else base_move_id
+        last_move_id = active_move_ids[-1] if active_move_ids else BASE_MOVE_ID
         return RuntimeLocation(RuntimeAddr(binary_loc), last_move_id)
     assert isinstance(runtime_loc, RuntimeLocation)
     return runtime_loc
@@ -185,7 +184,7 @@ def r2b(runtime_addr, specific_move_id=None, debug=False):
 
         # Early out if we have no relevant move ids
         if len(relevant_move_ids) == 0:
-            return BinaryAddr(int(runtime_addr)), base_move_id
+            return BinaryAddr(int(runtime_addr)), BASE_MOVE_ID
 
         specific_move_id = None
         if len(relevant_move_ids) == 1:
@@ -203,9 +202,9 @@ def r2b(runtime_addr, specific_move_id=None, debug=False):
             return (None, None)
 
     # Return the binary address for the runtime address and the selected move id.
-    move_dest, move_source, move_length = move_definitions[specific_move_id]
-    if move_dest <= runtime_addr <= (move_dest + move_length):
-        return (BinaryAddr(move_source + (runtime_addr - move_dest)), specific_move_id)
+    md = move_definitions[specific_move_id]
+    if md.is_in_move_dest(runtime_addr, include_end_address=True):
+        return (md.convert_runtime_to_binary_addr(runtime_addr), specific_move_id)
     return (None, None)
 
 # TODO: Maybe use this in more places
@@ -229,10 +228,10 @@ if __name__ == "__main__":
     id2 = add_move(0x70, 0x2000, 8)
     id3 = add_move(0x900, 0x2100, 256)
 
-    assert move_id_for_binary_addr[0x70] == base_move_id
+    assert move_id_for_binary_addr[0x70] == BASE_MOVE_ID
     assert move_id_for_binary_addr[0x1900] == id1
     assert move_id_for_binary_addr[0x2000] == id2
-    assert move_id_for_binary_addr[0x2000 + 8] == base_move_id
+    assert move_id_for_binary_addr[0x2000 + 8] == BASE_MOVE_ID
 
     assert b2r(0x70) == 0x70
     assert b2r(0x1900) == 0x70
@@ -247,16 +246,16 @@ if __name__ == "__main__":
         assert active_move_ids == [id2]
         assert r2b(0x70) == (0x2000, id2)
         assert r2b(0x900) == (0x2100, id3)
-        assert r2b(0x2008) == (0x2008, base_move_id)
+        assert r2b(0x2008) == (0x2008, BASE_MOVE_ID)
         with moved(id1):
             assert active_move_ids == [id2, id1]
             assert r2b(0x70) == (0x1900, id1)
             assert r2b(0x900) == (0x2100, id3)
-            assert r2b(0x2008) == (0x2008, base_move_id)
+            assert r2b(0x2008) == (0x2008, BASE_MOVE_ID)
         assert active_move_ids == [id2]
         assert r2b(0x70) == (0x2000, id2)
         assert r2b(0x900) == (0x2100, id3)
-        assert r2b(0x2008) == (0x2008, base_move_id)
+        assert r2b(0x2008) == (0x2008, BASE_MOVE_ID)
     assert active_move_ids == []
     assert r2b(0x70) == (None, None)
     assert r2b(0x900) == (0x2100, id3)
